[TV Control] Dropped level for parental control, profiling out of scope
There is no good reason to restrict parental control to the channel level,
even though the current API does that. A given channel may be for all audience
most of the time and broadcast content that is not suitable for children at
nights.

Also fine-tuned the ability to present audio/video streams: there may be more
than one selected stream to present. Profiling of the API for devices that can
only render a given number of simultaneous streams is explicitly flagged as
out-of-scope.