[TV Control] Accessibility feedback integrated
The APA WG is now explicitly mentioned in the list of groups with which the
TV Control WG will coordinate.

Examples of EPG information now also mention captioning and secondary audio
channel.

The Coordination section now includes a sentence to emphasize that the TV
Control WG cannot expect to meet the milestones at the expense of horizontal
reviews: horizontal groups need to be given sufficient time to review the spec.