[PNG-2021] Avoid citing an Editor's Draft as if it were the /TR
The title "Portable Network Graphics (PNG) Specification (Third Edition)" belongs to a document that will eventually live in /TR.  There is already a citation to its Editor's Draft shortly after this name is used.  Until the document exists in /TR, the title should not be associated with a draft in GitHub.