Avoid mentioning PR directly
There are proposed changes in the Process about simplifying the REC
track that if adopted would result in Proposed Recommendation being
retired as a maturity stage. The requirements to exit CR would remain
the same though.

This rephrasing recasts the criteria expect in order to enter PR in
terms of exiting CR instead. This is strictly equivalent, but avoids
mentioning PR by name, making this language more forward compatible with
this potential upcoming Process change.