Coordination: add SocialCG/IndieAuth, drop all rel=nofollow
Coordination / W3C Groups. Added Social Web Community Group in particular on identity topics that may overlap with the IndieAuth spec from that group / prior Social Web WG.

Also dropped all the rel="nofollow" attributes on links. 

We should not be using rel="nofollow" in W3C docs/charters etc. 

Certainly it makes no sense to do so on W3C to W3C links, nor even W3C to GitHub links. Also it seems rude to use rel="nofollow" on links to organizations and groups we expect to coordinate with. If they're good enough for us to coordinate with and link to, we should absolutely not be using rel="nofollow" (nevermind that any first party use of rel="nofollow" violates the intent of the rel-nofollow spec).